requirements.txt
customtkinter
Pillow
groq
python-dotenv
